requirements.txt
cv-bridge==1.13.0.post0
lz4==4.4.4
mcap==1.3.0
mcap-ros2-support==0.5.6
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
pillow==12.0.0
rosbags==0.11.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
tqdm==4.67.1
typing_extensions==4.15.0
zstandard==0.25.0
